+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- New accordion component added
+- New accordion and details components added
 
 ### Changed
 ### Deprecated